now the player can move where he suppose to

+++ includes/types.h
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   types.h                                            :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: omghazi <omghazi@student.42.fr>            +#+  +:+       +#+        */
+/*   By: hbettal <hbettal@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/07/30 20:35:41 by omghazi           #+#    #+#             */
-/*   Updated: 2024/08/15 13:44:05 by omghazi          ###   ########.fr       */
+/*   Updated: 2024/08/20 04:14:18 by hbettal          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -31,16 +31,16 @@ typedef struct s_parser t_parser;
 
 struct s_map
 {
-	char	**map;
-	int		largest_line;
-	int		height;
-	int		width;
-	void	*mlx;
-	void	*win;
-	void	*img;
-	t_player *player;
-	t_garbage *garbage;
-	t_parser *parser;
+	char		**map;
+	int			largest_line;
+	int			height;
+	int			width;
+	void		*mlx;
+	void		*win;
+	void		*img;
+	t_player	*player;
+	t_garbage	*garbage;
+	t_parser	*parser;
 };
 
 struct s_parser
@@ -62,8 +62,8 @@ struct s_player
 
 struct s_cordonnees
 {
-	int x;
-	int y;
+	double x;
+	double y;
 };
 
 struct s_garbage
